moved api req
