+++ acvillager/src/components/villagers/VillagersCollection.js
@@ -1,11 +1,36 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
+import axios from 'axios'
 import SingleVillagerCard from './SingleVillagerCard'
 
-function VillagersCollection({ pureVillagerArray, loading }) {
+const VillagersCollection = () => {
+
+    const [villagers, setVillagers] = useState([])
 
     const [searchVillager, setSearchVillager] = useState('')
 
-    function customFilter(character) {
+    useEffect(() => {
+        fetchVillagers();
+    }, [])
+
+    const fetchVillagers = () => {
+        // required headers for using Nookpedia API
+        const config = {
+            headers:
+            {
+                'X-API-KEY': `${process.env.REACT_APP_NOOKAPIKEY}`,
+                'Accept-Version': '1.0.0'
+            }
+        }
+
+        axios
+            .get(`https://api.nookipedia.com/villagers`, config)
+            .then(res =>
+                setVillagers(res.data),
+            )
+            .catch(error => console.log(error));
+    }
+
+    const customFilter = (character) => {
         if (searchVillager === "") {
             return character
         } else if (character.name.toLowerCase().includes(searchVillager.toLowerCase())) {
@@ -13,18 +38,14 @@ function VillagersCollection({ pureVillagerArray, loading }) {
         }
     }
 
-
-    if (loading) {
-        return <h3>loading...</h3>
-    }
-
     return (
         <div>
             <input type="text" placeholder="villager name..." name="villager-search" onChange={(event) => {
                 setSearchVillager(event.target.value)
             }} />
+
             <section className="villager-list">
-                {pureVillagerArray.filter((character) => {
+                {villagers.filter((character) => {
                     return customFilter(character)
                 }).map((villager, index) => (
                     <SingleVillagerCard key={index} villager={villager}>
@@ -32,6 +53,7 @@ function VillagersCollection({ pureVillagerArray, loading }) {
                 ))}
             </section>
         </div>
+
     )
 }
